refactor: move crypto to infra
